refactor(category): move selector to components directory
