+++ src/components/category-selector/CategorySelector.jsx
